add the ability to do something on key callback. Turns out this isnt what i wanted to do anyway, but i'll leave it there since it's still useful
oh boy was it mildly irritating to get that working
my first thought was to just pass the function as the callback, but
then it wouldnt accept python functions, so I made a wrapper with the
Window extension class, but that didnt work since `self` wasnt passed
(and since GLFWwindow* was the first argument it segfaulted). So now
there's a global dict mapping GLFWwindow* to Window extension classes.